Add tests when from value is 0x80

+++ Araki/JPL/ch20/ex20_01/TranslateByte.java
@@ -24,7 +24,7 @@ public class TranslateByte {
         int b;
         
         while ((b = in.read()) != -1) {
-            out.write(b == from ? to : b);
+            out.write((byte)b == from ? to : b);
         }
         
         out.flush();
